requirements.txt
numpy==2.3.5
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
pymongo==4.9.1
motor==3.6.0
pandas==2.2.3
langchain==0.3.7
langchain-openai==0.2.5
langgraph==0.2.35


